Now uses RootTools for su stuff
The app now "works" on the phone, but is completely useless. If I change
the MAC, then it doesn't seem to affect existing WIFI connections. And
stopping and restarting WIFI reverts the MAC to the original value.

I grepped for the MAC address string, and found a file at
/data/misc/wifi/WCNSS_qcom_cfg which contains it without any :

Next step is to backup then edit that file.

Something like:

1) Stop WIFI.
2) Backup cfg file.
3) Edit cfg file.
4) Start WIFI.